# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "version": "1.0.0"
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "consumes": [
        "multipart/form-data"
    ],
    "paths": {
...
        "/api/listesEtudiantsMerites": {
            "get": {
                "summary": "Lister Les étudiants (par mérite)",
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Etudiants par mérites"
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {}
                            },
                            "example": {}
                        }
                    }
                }
            }
        },
        "/api/ajoutEtudiant/Merite": {
            "post": {
                "summary": "Ajouter un Etudiant",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorize",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Etudiants par mérites"
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "roles_id": {
                                        "type": "string"
                                    },
                                    "photo_profile": {
                                        "type": "string"
                                    },
                                    "INE": {
                                        "type": "string"
                                    },
                                    "date_naissance": {
                                        "type": "string"
                                    },
                                    "lieu_naissance": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "sexe": {
                                        "type": "string"
                                    },
                                    "niveau_etudes": {
                                        "type": "string"
                                    },
                                    "filiere": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "statuts_id": {
                                        "type": "string"
                                    },
                                    "moyennes": {
                                        "type": "string"
                                    }
                                }
                            },
                            "example": {
                                "prenom": "Astou ",
                                "nom": "Diouf",
                                "email": "astou849@gmail.com",
                                "password": "Pass123@",
                                "roles_id": "4",
                                "photo_profile": "image.jpg",
                                "INE": "N01662320224",
                                "date_naissance": "12/01/1999",
                                "lieu_naissance": "Dakar",
                                "adresse": "Dakar",
                                "sexe": "Homme",
                                "niveau_etudes": "Licence 3",
                                "filiere": "Lettres Modernes",
                                "telephone": "771334567",
                                "statuts_id": "1",
                                "moyennes": "18.00"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/listesEtudiantsCasSocial": {
            "get": {
                "summary": "Lister les étudiants (par cas social)",
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Etudiants par cas social"
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            },
                            "example": {
                                "email": "etudiant2@gmail.com",
                                "password": "Etudiant1@"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/ajoutProfil": {
            "post": {
                "summary": "Ajouter un User (Profil)",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorize",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "User Chef de Service Pédagogique"
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "roles_id": {
                                        "type": "string"
                                    },
                                    "photo_profile": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "sexe": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    }
                                }
                            },
                            "example": {
                                "prenom": "ChefPedaogique 1",
                                "nom": "UCAD",
                                "email": "ChefPedaogique1@gmail.com",
                                "password": "ChefPedaogique1@",
                                "roles_id": "3",
                                "photo_profile": {
                                    "type": "string",
                                    "format": "binary"
                                },
                                "sexe": "Homme",
                                "telephone": "770000000"
                            }
                        }
                    }
                }
            }
        },
...
    }
}